Flytter felleslogikken som avviser utløpte tokens.
