+++ src/main/kotlin/no/nav/personbruker/dittnav/api/health/authenticationCheck.kt
@@ -5,7 +5,7 @@ import io.ktor.http.ContentType
 import io.ktor.response.respondText
 import io.ktor.routing.Route
 import io.ktor.routing.get
-import no.nav.personbruker.dittnav.api.legacy.executeOnUnexpiredTokensOnly
+import no.nav.personbruker.dittnav.api.config.executeOnUnexpiredTokensOnly
 
 fun Route.authenticationCheck() {
